requirements.txt
requests
beautifulsoup4
python-telegram-bot==22.1
